fix : Service Interface 제거
