+++ docs/jihoon/요구사항 체크리스트.md
@@ -42,6 +42,7 @@
 - [ ] @Valid 달 부분 없는지 Refactor
 - [ ] ProductQueryOrderFactory 로직 개선
   > Product의 컬럼에 대한 Sort값이 들어오지 않는 경우 예외처리
+- [ ] 관심상품 > userDetails 를 통해 userService를 사용하여 user를 가져와야함
 
 # Test List
