+++ src/main/java/bc1/gream/domain/order/controller/OrderCommandController.java
@@ -1,0 +1,5 @@
+package bc1.gream.domain.order.controller;
+
+public class OrderCommandController {
+
+}
