+++ src/main/java/bc1/gream/domain/order/controller/OrderQueryController.java
@@ -1,4 +1,4 @@
-package bc1.gream.domain.order.controller.query;
+package bc1.gream.domain.order.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
