--- src/main/java/bc1/gream/domain/order/controller/command/OrderCommandController.java
@@ -1,5 +1,0 @@
-package bc1.gream.domain.order.controller.command;
-
-public class OrderCommandController {
-
-}
