+++ src/main/java/bc1/gream/domain/order/service/query/OrderQueryService.java
@@ -1,10 +1,19 @@
 package bc1.gream.domain.order.service.query;
 
 import bc1.gream.domain.order.entity.Order;
+import bc1.gream.domain.order.repository.OrderRepository;
 import bc1.gream.domain.product.entity.Product;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface OrderQueryService {
+@Service
+@RequiredArgsConstructor
+public class OrderQueryService {
 
-    List<Order> findAllTradesOf(Product product);
+    private final OrderRepository orderRepository;
+
+    public List<Order> findAllTradesOf(Product product) {
+        return orderRepository.findAllByProduct(product);
+    }
 }
